New SongAutocomplete replaces AutocompleteInput and reduces code base in PicksForm
